Backend/requirements.txt
asgiref==3.11.1
Django==6.0.3
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.11
PyJWT==2.12.1
razorpay==1.4.2
python-decouple==3.8
setuptools==80.9.0
sqlparse==0.5.5
tzdata==2025.3
